refactoring common code out of qtconsole
